test(refactor)

+++ test.sh
@@ -4,24 +4,23 @@ export $(cat .env-dev | xargs)
 
 LDFLAGS="-X github.com/wearedevx/keystone/pkg/client.ksauthURL=$KSAUTH_URL -X github.com/wearedevx/keystone/pkg/client.ksapiURL=$KSAPI_URL"
 
-#go test -parallel 1 -tags test -ldflags "$LDFLAGS" -work  "$@"
-
 # Create db file
 touch $TMPDIR/keystone_gorm.db
 
 # Start test
-go test -tags test -ldflags "$LDFLAGS" -work  "$@"
+go test -tags test -ldflags "$LDFLAGS" -work "$@"
 
 ksauthpidpath=${TMPDIR}keystone_ksauth.pid
 
-pid=`cat $ksauthpidpath`
+# Check if gcloud auth func pid exist
+if [ -f "$ksauthpidpath" ]; then
 
-# Stop gcloud function
-# echo "kill -- -$(ps -o pgid=$pid | grep -o '[0-9]*')"
-# pkill -P $pid
-rm $ksauthpidpath
+    pid=$(cat $ksauthpidpath)
 
-kill -- -$pid
+    rm $ksauthpidpath
+
+    kill -- -$pid
+fi
 
 # Delete db file
 rm $TMPDIR/keystone_gorm.db
